[wasm] Skip WasmEngine::EstimateCurrentMemoryConsumption() if unused
Same as for https://crrev.com/c/5757568, skip the memory consumption
estimation if the result is not used (as the histogram is disabled).

This should fix the remaining flakes for the many-modules mjsunit test
in combination with TSAN and stress_concurrent_allocation.

Bug: 349610478, 355462536
Change-Id: Ia093e95d289522c1ba3c0b154dcef189fa362057
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5777002
Reviewed-by: Eva Herencsárová <evih@chromium.org>
Commit-Queue: Eva Herencsárová <evih@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95554}