[interpreter,parser] Move async function AST rewrite into bytecode generator
Due to historical reasons, some features are done via AST rewrites in
the parser, and some are done in the bytecode generator. In general,
doing it in the bytecode generator is more understandable and simpler.

Async functions have an AST rewrite that surrounds the entire body with
a try-catch so that synchronous errors can be turned into Promise
rejections.

This AST rewrite is undesirable complexity and also very difficult to
correctly nest with e.g. the try-finally needed for the explicit
resource management proposal.

Also drive-by refactors BytecodeGenerator methods that wrap the function
body with special logic (e.g. async functions' try-catch).

Bug: 358404372
Change-Id: I1e5ad6e6eafa576c58a99b22c58388b5d93fa944
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5763647
Reviewed-by: Rezvan Mahdavi Hezaveh <rezvan@chromium.org>
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95567}