[parser] Reduce crashiness of bad merges
There is a bug around compilation merging (streaming
compilation/deserialization merging with existing isolate cached code)
where we should be using the old script/shared function infos, but are
using the new one. This is not what we want, but it's not the end of the
world if it happens as long as the old and new objects are compatible.

So, while we investigate this issue, reduce the number of hard crashes
by using DumpWithoutCrashing, and making the new SFI/Script setup a bit
more robust.

Also print more useful parameters in the non-crashing dumps.

Bug: 355575275
Change-Id: I238ac5d89e2c96967b1cca9c25c2f8f0486faf07
Fixes: 358665167
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5783233
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95582}