[memory64] Mitigate SciMark2.1 regression of wasm64 trap handling
The new trap-handling-based bounds checks cause a regression in a few
of the v8/SciMark2.1 benchmarks. In particular, "Sparse matmult"
regressed 16% with LLVM-wasm64-turboshaft. Part of the regression is
caused by the fact that with trap handling we introduce ProtectedLoads
and ProtectedStores in place of simple Loads and stores, which causes
some suboptimal code generation. For example:

   vmulsd xmm3,xmm3,[r8+r12*1+0x10]

may become:

   vmovsd xmm4,[r8+r12*1+0x10]
   vmulsd xmm3,xmm3,xmm4

This CL addresses this issue, allowing Protected loads and stores to be
eliminateable instruction, moving the "protected" access mode to the
float binop instruction that embeds the load. This should be useful with
code that does floating-point matrix operations, for example, and also
works for 32-bit memories. In fact we measure the following improvements
in SciMark2.1

Windows/x64
Sparse - 64 bit: 1052 -> 1126
    LU - 64 bit: 1930 -> 2199
Sparse - 32 bit: 2176 -> 2334
    LU - 32 bit: 3588 -> 3676

Linux/x64
Sparse - 64 bit: 1369 -> 1515
    LU - 64 bit: 3183 -> 3469
Sparse - 32 bit: 2294 -> 2677
    LU - 32 bit: 3846 -> 3971

This should be extended to integer binop operation and to arm64.

Bug: 345936778
Change-Id: Ib203f0379f1cf494e5bcf3b3dedc4576c95bf92b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5666630
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#95595}