[maglev] Don't change object that other objects depends on
If we have two objects A and B, such that A points to B (it contains
B in one of its field), we cannot change B without also changing A,
even if both can be elided. For now, we escape both objects instead.

Fixed: 356913463
Change-Id: I297e6a40fb38201bc71551f2b3342ce8457ded73
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5783682
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95605}