[interpreter][sparkplug] Fix deopt loop for exceptional comparisons
When performing relational comparisons with a symbol, e.g.
`2 < new Symbol("test")`, an exception is triggered. This however
skips writing back the feedback meaning that when turbofan tiers up,
it still doesn't have any type feedback for the operation and emits an
unconditional deopt causing a deopt loop.

This CL fixes that by adding a new block to these built-ins which
writes back the feedback when an exception is thrown.

Change-Id: Iad2fe9e227cabd74fcefd4a649ab261dd23ee137
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5783528
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95633}