[cfi] Introduce a SegmentedTable
and use it as a base class for the ExternalEntityTable.
The SegmentedTable class moves out the generic functionality of the
ExternalEntityTable (freelists, segments) so that it can be used for the
Wasm code pointer table proposed in: go/v8-cfi:wasm-forward-edge

This first CL mostly moves over the functions except that I moved the
freelist initialization from ExternalEntityTable::Extend() to
SegmentedTable::AllocateTableSegment() and
ExternalEntityTable::AttachSpaceToReadOnlySegment().

Bug: 42203297
Change-Id: Id4570826f5324ff86ae01c9b1a341a65e3a930c6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5782818
Commit-Queue: Stephen Röttger <sroettger@google.com>
Reviewed-by: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95624}