[deoptimizer][wasm] Mark heap inaccessible during deopt
As the deoptimizer reads and writes "arbitrary" stack slots based on
deoptimization data, it is important for the V8 sandbox that none of
these actions are based on any on-heap data (which could be invalid
in case of an in-sandbox-corruption) as that would provide a sandbox-
escape.

As wasm deoptimizations should not need to access the heap at all,
simply install a scope for wasm deoptimizations blocking heap accesses
on modifiable heap pages for the duration of the deoptimization
completely.

Bug: 42204618
Change-Id: Ifc090e30e42f8c021a39dd241de8eca6d4753635
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5782746
Reviewed-by: Samuel Groß <saelo@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95638}