[builtins-pgo] Use meta.json instead of profile_version
Previously, we generated a profile_version file to store the currently
downloaded version, and skipped the meta.json download. Going forward,
we need to download the meta.json to retrieve the applicable V8 revision.

This change adds the meta.json file to the downloader and compares the
currently downloaded version with the details in meta.json instead of
the profile_version.

Additionally, we wipe the profile directory to avoid clashes when
switching back and forth between V8 releases as a developer or a
release builder.

Bug: 353262752
Change-Id: I5845620667c6753ff4ed266d681ca0fd2ba6cd3e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5782831
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Alexander Schulze <alexschulze@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95654}