[sandbox][leaptiering] Basic serializer support for dispatch handles
This CL implements basic support for serializing and deserializing
JSDispatchHandles. Previously, serializing an object would clear any
dispatch handles in the object. However, now that builtin JSFunctions
have a dispatch entry, the serializer needs to support them. With this
CL, the deserializer will now allocate entries in the JSDispatchTable
with the correct parameter count. Serializing the Code object stored in
the table entry will be done in a later CL.

Bug: 40931165, 42204201
Change-Id: I35b7eb21efc6eab74dc0be305a8b5f5c8c0069ed
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5783522
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95685}