[tiering] Apply PGO to JSFunctions with empty feedback vectors on creation
The problem is that the code for
profile_guided_optimization_for_empty_feedback_vector is not executed
after feedback vector creation.
This can be reproduced in scenarios like:
  JSFunctionBuilder::Build()
    JSFunction::EnsureFeedbackVector()
    Compiler::PostInstantiation()
      JSFunction::InitializeFeedbackCell()
        if (function->has_feedback_vector())
          return;
        profile_guided_optimization_for_empty_feedback_vector
The solution is to apply profile-guided optimization to all JSFunction
instances with empty feedback vectors as soon as the vector is created.
This results in an increased number of right now compiled functions when
running Speedometer3:
  Maglev compilation: ~26% increase (590 -> 743)
  Turbofan compilation: ~56% increase (90 -> 141)

Bug: v8:14296
Change-Id: I7d53e10c3662577b016a7c1671739a1d56152e7e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5790485
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Tao Pan <tao.pan@intel.com>
Cr-Commit-Position: refs/heads/main@{#95736}