[fastapi] Use fast API calls only when the number of parameters matches
So far the function overload resolution of fast API calls worked such
that the API function got called with N parameters, then the call target
that takes N parameters will get called. However, if N is higher than
the parameter count that any of the call targets needs, then the target
with the highest parameter count was called.

However, this overload resolution caused problems when not all overloads
of an API function were marked as fast-api compatible.

With this CL, the fast API will only be used when the number of
parameters passed to a function call matches the expected number of
parameters for at least one of the API function overloads.

Bug: 40259610
Change-Id: I5382351726aeb1b4b9f6e8f4fd9bbe4423a738d0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5797463
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95705}