[sandbox][leapriering] JSFunction::set_code updates JSDispatchTable
With this CL, when JSFunction::set_code is called, it now also updates
the Code object referenced from its entry in the JSDispatchTable. This
requires a few more miscellaneous changes to ensure that the parameter
count of the entry and of the Code object match (otherwise, there would
be a CHECK failure when setting the Code object).

Bug: 40931165, 42204201
Change-Id: I0786a55ca4e04d680a4f55e9aa20c268eac5bfed
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5785964
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95702}