[shadow stack] Support lazy deoptimization
Once a code is marked for deoptimization, V8 iterates stack and replace return
addresses of deoptimized functions to trampoline for deoptimization, which
violates CET policy. Instead, this CL patches the generated code to jump to the
trampoline after returned.

The generated code will look like this when it's not marked for deopt:
call function
nop (5-bytes)

And the code is patched to the following when marked for deopt.
call function
jmp deopt_entry

Bug: 42201233
Change-Id: Id6f22859cc7c1992f388b2945071338144188e45
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5726440
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Choongwoo Han <choongwoo.han@microsoft.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95746}