[cfi] memory sealing for the global code range
Add a new PageAllocator::SealPages function and use it to seal the
process-wide CodeRange to prevent attacks outlined in the last
paragraph of https://v8.dev/blog/control-flow-integrity#corrupted-syscall-arguments.

The feature is added behind a gn arg since we don't have support on our
bots for this yet. Once the mseal syscall is available on our bot, we
can enable this by default.

Discarding sealed memory requires write-access to the page. For this,
there's a new DiscardSealedMemoryScope to temporarily acquire write
permissions if the JIT is pkey-tagged.

Bug: 360048056
Change-Id: I1115c1b16c7b752725761ede8f0a9261b71c787e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5791010
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#95706}