[regexp] Improve analysis around positive lookaround.
In https://issues.chromium.org/issues/42201279 a bug
was discovered with eats_at_least and lookarounds.
The fix disabled the eats_at_least analysis around
positive lookarounds.  This change partially restores
the optimization, ignoring the contents of the lookaround,
but using the eats_at_least data from whatever comes after
it.

For example, the regexp /(?=<")hest[-_]fisk(?=")/ was not
able to use the Boyer-Moore optimization because the lookbehind
destroyed the eats-at-least information. With this change it
can use Boyer-Moore and is 20x faster.

One tricky aspect of this change is that if a node claims to eat
at least n characters it must fill in Boyer-Moore data for that
number of characters.  These parts of the code are not near each
other and it is easy to get this wrong.

Bug: 42201836
Change-Id: I019b0deb6db3c5fea77ecd307e609c4500cb20df
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5676708
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95739}