[wasm] Refactor central stack switches
Instead of storing the stack limit and stack pointer in a fixed slot of
the switching frame, store this information in the StackMemory. This
makes the logic independent of the frame type, which will make it easier
to extend to other code kinds, and in particular to wasm builtins.

Bug: 349640002,42202153
Change-Id: Iac523b53d0c4e0f184dc7ac736d3a41ecb0b7228
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5797181
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95742}