[turboshaft][wasm] Use TS variables to handle InstanceCache
The InstanceCache (named like that because its main usage was to cache
commonly used fields of the WasmTrustedInstanceData) caches two
mutable values:
- memory start of memory 0
- memory size of memory 0

These values can be impacted by operations like memory.grow, so the
graph builder needs to be careful to handle the cached values across
all control flow edges which turns out to be quite error-prone.

This CL removes the manual handling of merges across control flow and
relies on Turboshaft's `Variable` feature which does almost all of
that automatically. (The one exception being loops without a backedge,
in that case the wasm graph builder still needs to replace the
PendingLoopPhiOps with regular PhiOps (which is done easily as we
already do that for all the other PendingLoopPhiOps for the values of
the wasm value stack.))

Bug: v8:14108
Change-Id: I1d7b51f2061c5264ed98ebf1c46b20d2cd19a080
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5803584
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95751}