[wasm][jspi] Switch to the central stack for JS builtins
Switch to the central stack when calling a JS builtin from a wasm
builtin, otherwise the JS builtin can perform arbitrary actions that may
be unsafe to perform on a secondary stack such as calling a runtime
function.
We may be able to allowlist some of these calls in the future for
builtins that are known to be safe.

R=jkummerow@chromium.org

Bug: 349640002,42202153
Change-Id: I35b6c1739bbbb3c5f09b37f2a4163e5c7256c95d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5804144
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95771}