[x64][wasm] Fix wrong source position on fused load + float binop
https://crrev.com/c/5666630 introduced combining a protected load
(meaning a load which potentially needs to trap but is implemented
using the signal handler based on memory guard pages for linear memory
or the WasmNull for null-dereferences) and a floating point binary
operation (like the f32.add in the reproducer) into a single machine
code operation.

However, the optimization did not update the source position leading
to a wrong error location and for wasm-gc instructions even a wrong
error message (as it treated the failing load as a "memory out of
bounds" while it should have been a "dereferencing a null pointer".

Bug: 345936778
Fixed: 361611472
Change-Id: I9f893a6aebebfb73aaa4467a7918faf51286ab3c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5812143
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95812}