[heap] Fix race in detaching cppheap
In case the heap owns the cppheap, the cppheap is currently detached as
soon as we start tearing down the heap. At that time  background
compilation threads may still be active and may try to allocate, which
means they will check allocation limits and try to access the cppheap to
read the embedder size. This results in a race between detaching the
cppheap and the background compiler.

Fix by delaying detaching the cppheap such that it happens after the
background threads have been stopped.

Bug: 361711106
Change-Id: I984609026d788497910be8f5ed8f1762bffaec51
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5816135
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95839}