[sandbox][leaptiering] Context specialized dispatch entries
Introduce a new bottleneck `UpdateCode` for changing the code of
JSFunctions. This one ensures a context specialized js function has its
own context specialized dispatch entry.

When a function is specialized we create a dedicated entry and re-wire
it to use that one instead.
When a function is un-specialized we merge it back into the canonical
handle that we can get from the feedback cell.

This is an intermediate state. Only dispatch entries are ensured to have
the correct properties. The code inside the dispatch entries can still
be out-of-date, as we are not yet updating them in CompileLazy and the
interpreter entry stub.

Bug: 40931165, 42204201
Change-Id: I400e82e61cc4e34e2fa28cba63344fb350953e55
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5803648
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95848}