[sandbox][leaptiering] Call JSFunctions through the JSDispatchTable
With this CL, JSFunctions are now invoked through the JSDispatchTable.
We still use the parameter count from the SharedFunctionInfo. This will
be addressed in a follow-up CL.

With that, we can now also remove the JSFunction::code field when
leaptiering is enabled as it is then no longer in use.

Bug: 40931165, 42204201
Change-Id: Ic7bfe353521e7d894e3550d926fe9ba0cbbc6e78
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5803650
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95870}