[build] Enable logging mksnapshot output when verifying determinism
This adds a feature to run.py that enables logging stdout to a file. The script is used as a wrapper to all non-python gn actions.

We use it now in mksnapshot to log output on the determinism bots with some tracing turned on (--trace-read-only-promotion).

The logs are treated as data and as such get isolated to CAS after compilation for debugging purposes.

Bug: 338419287
Change-Id: Iba426d0b012fd4eb9c7c8e2c7dffae46aad11248
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5826652
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95883}