[wasm] Skip WasmJSFunctions in js-to-wasm wrapper tier-up
When a js-to-wasm wrapper tiers up, we also set the newly compiled
wrapper as the target for other exports that have the same signature.
This assumed that all exports have type WasmExportedFunction, but they
can also have type WasmJSFunction in the case of a re-exported
WebAssembly.Function import.

R=clemensb@chromium.org

Fixed: 362539773
Change-Id: I190a680ac5726122e2124977668bba3a95df93b5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5822928
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95877}