mksnapshot: Improve determinism for RO space promotion
Read-only (RO) space promotion is a mechanism to copy islands of
graphs to RO space during snapshot creation. The goal is to promote
objects to RO space for performance reasons and maintain a
deterministic RO space layout while doing so.

The previous approach used std::sort() on a vector of collected
promotion candidates. The sort was based on absolute addresses trying
to create a stable RO space layout. It assumed that Heap and its
backing allocators are able to always acquire the same pages wrt to
absolute addresses from the OS. This is not always the case, even with
--predictable.

However, pages are maintained in a linked list in Heap after they are
acquired from the OS. Using HeapObjectIterator yields in a stable
order. The CL maintains insertion order using a separate vector
instead of relying on sorting based on addresses.

We also pass --predictable to mksnapshot which results in GCs that
don't reorder pages as a result of using concurrent helper tasks.

Bug: 338419287
Change-Id: I00cd45fe91bf38e90024bc99ee474daf791519d9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5823375
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95885}