[wasm] wrapper cache: speed up table maintenance
Asking the global wrapper cache whether a given call target belongs
to a known wrapper can be relatively slow, because it causes mutex
contention in applications with many worker threads. So this patch
adds thread-local lookup caches to minimize the number of requests
to the mutex-protected global wrapper cache.
Bonus change: a RecursiveMutex on the WasmImportWrapperCache no
longer needs to be recursive, so it is replaced with a plain Mutex.

Bug: 362535231
Change-Id: I24f5d8c2fcc19b4ff440e67eb7d2b1053662f99a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5824989
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95882}