[maglev] Ensure VO phis are set for all inputs
We have an optimization that avoids merging fields when two
equal VirtualObjects flow into the same block. However, this
approach can lead to issues when one of the fields is a phi node.

Consider a scenario with a VO phi that has three predecessors.
The first two predecessors may differ, prompting a merge. If
the third predecessor is similar to the last one merged, the
merge is skipped. However, if the third predecessor differs
from the last merged one (even if it's the same as a previous
one), we attempt to merge again.

This change ensures that all VO phi inputs are properly set.

Fixed: 362122733
Change-Id: I948e5d4ead5e2ad2cce63d483b853d2ead13d706
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5833381
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95922}