[maglev] Add quick check for RO values to write barrier
With static roots we can use a simple comparison with an immediate
to check if a value is in RO space. As this is very cheap, we add
this check before loading the page flags.

Bug: 364466897
Change-Id: Ife1f596d6b3de2d28d4bdc76561fe2ba54cf0bb8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5833498
Auto-Submit: Patrick Thier <pthier@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95942}