[sandbox][leaptiering] Refactor and simplify some dispatch handle code
Currently, it is possible to initialize a dispatch handle after the
object has already been created. This operation may require a write
barrier, depending on whether the object could've been marked in the
meantime. With this CL, we now always perform a write barrier when
initializing the JSDispatchHandle to be conservative.

To simplify the code and make it easier to reason about, this CL also:
* Drops unused dispatch handle initialization methods
* Drops some unnecessary parameters from the dispatch handle
  initialization methods that are in use
* Simplifies the write barrier logic in JSFunction::UpdateCode
* Adds DCHECKs to ensure that we never overwrite an existing dispatch
  handle in a FeedbackCell (the handle there is const-after-init)

Bug: 40931165, 42204201, 364555672
Change-Id: I4236a7b35b71ac2a7bd3cd50007b179651029e06
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5835723
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95949}