[wasm] Remove unneeded check in js-to-wasm wrapper tierup
We should never trigger tier-up of the start function, and if we do we
should fix it by adjusting the tiering logic.

Currently nothing fails, so this CL removes the unneeded code and adds a
comment in case this fails with some weird flags or on some fuzzer one
day.

R=ahaas@chromium.org
CC=dlehmann@chromium.org

Bug: 363916147
Change-Id: Ic80d12928b1597519aade9ac682fdb86aae5a186
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5836882
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95971}