[promises, async stack traces] Fix DCHECK failure caused by async task Id overlow
Currently, async task Id value in V8 is limited to 22 bits which can
easily overflow by a long running application. This CL makes it
explicit that the async task Id is wrapped around max value to avoid
DCHECK failures.

With this workaround the same Id might be assigned to multiple tasks
which might cause an ancient async task to show up in a stack trace
instead of a correct one in DevTools. Given that it's an unlikely case
we consider it better than hitting a DCHECK.

Drive-by fixes:
 - avoid UB caused by overflowing of async task count value,
 - ensure that invalid task Id (zero) is never set.

Bug: 340606663
Change-Id: I5a525a7cd083618d69d0015fc51eca23c76e716f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5839866
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95978}