[d8] Handle fast API targets without signature correctly
In d8 there are several fast API call targets for tests. Most of these
targets have what is called a signature, which means in this context
that the type of the receiver has to be a specific API object. However,
some targets don't have a signature, and these targets can then be
called with an arbitrary receiver.

The API call targets that did not have a signature still accessed the
receiver, which caused crashes when the receiver was not of the expected
type, e.g. when the receiver was a Smi.

With this CL we try to unwrap the receiver in these API call targets,
but use a generic receiver object if the incoming receiver has an
invalid type.

R=ishell@chromium.org

Bug: 363983044
Change-Id: I6115af1cf742f7a76d13a7cf5cae6aa7b40168bb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5837111
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95980}