[fastapi] Handle returns without type conversion to tagged
So far, returns of fast API calls got converted unconditionally to
tagged. This meant that if the return type was float64, and the calling
function could handle a float64, there would still be a HeapNumber
allocation.

With this CL, the result of the fast API call is marked with the correct
type, so that e.g. float64 returns only get converted to HeapNumbers
when necessary.

Bug: 41497260
Change-Id: Ib027e2dbacf5ac12baba6f64d7bd45f99c5b95d2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5832361
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96032}