Make implicit `this` capture explicit
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in C++20.
It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This CL makes the implicit captureexplicit, preventing the warning.
It does not change the compiled code at all, since it's just removing
some syntactic sugar.

Bug: 351004963
Change-Id: I38cbc11eb9c7130f8094be5c6c008b932fd2dc0a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5854443
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#96078}