[fuzzer] Don't allow --fast-api-allow-float-in-sim in fuzzer runs
The flag was introduced to allow to test the fast API also in simulator
runs on specially prepared functions. However, in general, float
parameters don't work for fast API calls in simulator runs, because the
simulator does not know the signature of the called C function, and it
does not know which registers hold the parameters, and in which order.

R=machenbach@chromium.org

Bug: 366343787
Change-Id: I2138da10823d15c6456fc65fb8387eaec8d0e777
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5865731
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96294}