[fastapi][debugging] Add code comment
The implementation of fast API calls has a backup regular API call
implementation that is used when a parameter is invalid for the fast API
call. This regular API call is not supposed to return, as it is supposed
to throw an exception because of the invalid parameter, and if the
regular API call returns, it runs directly into an UNREACHABLE. This CL
adds a code comment at the return location of the regular API call so
that it's easier to debug why the UNREACHABLE was hit.

Change-Id: I5db56e0250445900bf371ed6e72996b1ca673280
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5870229
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96132}