Fix unreachable code in SourceTextModule::PrepareInstantiate()
When Wasm is disabled, there is a break statement that is unreachable,
which in turn triggers a -Wunreachable-code-break warning that then
becomes fatal due to -Werror. Fix this issue by moving the break
statement to the part of the code where Wasm is enabled.

Change-Id: I17ed2f3675c21485b348062ff35d4e4c47c0e67f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5870923
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Adam Klein <adamk@chromium.org>
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96149}