[fuzzing] Fix %WasmStruct() / %WasmArray() in differential fuzzing
when one of the targets uses --jitless.
This is a follow-up to https://crrev.com/c/5870230 which fixed the
issue of fuzzer crashes in differential fuzzing but introduced
differential fuzzers reporting differences between jitless and
non-jitless modes (due to making %WasmStruct() not doing anything at
all if jitting is disabled).

This change changes the way wasm objects are created, so that no wasm
function needs to be compiled.
Still, when using --jitless, instantiating a wasm NativeModule doesn't
work as it already reserves WasmCodeCache and performs other actions
that are not supported.
Instead, the builtins now return a frozen empty JS object which should
hopefully be consistent with how wasm objects behave in JavaScript.

Bug: 366323504
Fixed: 367734948
Change-Id: I8f8397a35f1eec408d0ad273a9a5836f55121e47
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5873559
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96164}