[heap] Moving growing in the GC epiologue for Scavenger
Growing was already in the epilogue for Minor MS and full GC. Adjust
Scavenger to also grow after the GC.  This allows for adjusting
growing heuristics to be based on survival rates.

This does slightly affect pretenuring for Scavenger only as we are now
computing heuristics based on the size that was used for the GC and
not the already larger new space. While this is more correct (in terms
of survival rates), it may move benchmarks. None of the interesting
benchmarks seem to move though.

Bug: 365027679
Change-Id: I4765e2ebfb685b5b76aa84089b2a6314fd9ec1bb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5865418
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96182}