Disallow Oilpan GC during heap snapshot generation
We have encountered a bug where NameInHeapSnapshot modified the heap,
which caused the heap snapshot generator to fail. This change adds a
disallow scope so that there would be a DCHECK failure if this bug
regressed, and also updates a comment to describe the expectation.

Bug: 367037118
Change-Id: Ic39043be5b76ff062264036a14c8a6315dc2a083
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5874863
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#96186}