[heap] Perform global safepoint in CollectCodeStatistics()
When using a shared heap Heap::CollectCodeStatistics() needs to
perform a global safepoint on the shared space isolate for
Heap::MakeHeapIterable().

This CL also now asserts that the clients mutex is locked in
IterateClientIsolates(). This uncovered more similar issues that all
need a global safepoint for shared space isolates. Therefore this CL
also adds a new ctor to SafepointScope with this exact behavior. This
ctor can be used when appropriate throughout the codebase without
duplicating logic.

Bug: 366635350
Change-Id: I4036be6d2e8d974d49d6086bf51ddf83c103cdb8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5872562
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96175}