[foozzie] Block sandbox-testing mode for differential fuzzing
Sandbox testing leads to false positives with differential fuzzing.

E.g. when we compare cross architecture, we rely on crashing on certain range errors. Sandobox testing seems to alter the crash in a way that leads to an output difference in https://crbug.com/366307030.

In https://crbug.com/368086266 we see a different behavior of a function that we neutralized for differential fuzzing: %GetOptimizationStatus.

Another case analyzed in http://b/367909533#comment2 it cuts off or alters the output when we encounter a crash. The resulting output difference is flagged by differential fuzzing.

No-Try: true
Bug: 367909533, 366307030, 368086266
Change-Id: Ie95ef891e5c1f8942d2fc8dcbf817905206f8a07
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5872498
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96181}