Separate out the option to compile the wasm fuzzer.
Previously all non-official builds had the wasm fuzzer.
This is still true by default, but now there's an
explicit build option to override that.

This is useful for anyone who wants to build the official
build with the fuzzing support, or anyone who wants to embed
a non-official build of V8, but doesn't want to pay for the
binary size increase of the fuzzing support.

Change-Id: I0617129940ab040dce9eb1a2ebcbb4f92421cf09
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5873224
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96231}