Silence a spurious error in MergeAssumptionChecker
When MergeAssumptionChecker runs on an eval script, it crashes due to an
unexpected pointer from the Script to its eval_from_shared. This pointer
is harmless; it is not part of the object graph built by the
compilation. Instead, it represents the context in which the compilation
took place.

Bug: 366635355
Change-Id: I86ac037234fe361d778aff1068db5239510828e7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5876479
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#96213}