[debug] Introduce a dedicated `StackTraceInfo`.
This CL lays the ground-work for tracking the correct async continuation
for each (synchronous) stack trace captured within V8 by adding a
dedicated `StackTraceInfo` internal representation of `v8::StackTrace`.
At this point, the CL is basically a no-op and only adds one layer of
abstraction in between; but in a follow-up CL we will hook up the async
parent information to the `StackTraceInfo`.

Bug: 350426235
Change-Id: Iab1f2b4001518e7dd9e3b317475ab6280a9327b3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5872849
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96199}