[turbofan] Add bailout for byteOffset for RAB/GSAB
byteOffset needs to check for whether the underlying buffer was resized
in a way that moves the offset out of bounds. For now, just bail out of
optimizing byteOffset for resizable elmenets; in the future we probably
want to optimize this same as we optimize length and byteLength.

Bug: v8:11111
Change-Id: I22a32f95be93bec3c4feda5dcf6f4e16b26d96b8
Fixed: 368076858
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5874051
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96221}