[wasm] Do not apply wasm_debug_mask_for_testing during deopt
When running in debug mode, we do not tier up. If we are still running
optimized code and this code triggers a deoptimization, we specifically
compile the unoptimized frame with kNotForDebugging (as otherwise the
deoptimizer would have to handle all special cases of debug side tables
etc.).

The --wasm-debug-mask-for-testing flag overwrote this kNotForDebugging,
forcing debug mode for the deopted function.

Bug: 42204618, 366350772
Change-Id: I1dfaaed72745c97e45223b6ec9f39592a40c5537
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5873251
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Commit-Queue: Manos Koukoutos <manoskouk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96248}