[objects] Fix usage of wrong isolate in worker isolate
In this crash JSReceiver::DeletePropertyOrElement was trying to load
the isolate of a SharedArray using GetIsolate(). However, this always
returns the shared space isolate in case of shared objects even on
the worker isolate. This led to a crash soon afterwards where the
worker thread was using the shared space/main isolate.

This CL fixes this by passing in the isolate as an argument to the
DeleteProperty, DeleteElement and DeletePropertyOrElement methods in
JSReceiver. This aligns well with our goal of getting rid of uses
of GetIsolateFromWritableObject().

Bug: 368060007, 336738728
Change-Id: I4212f6c1132480dbf900ee125592c8b86b73f261
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5881777
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96235}